+++ packages/g-library/rollup.config.ts
@@ -38,10 +38,12 @@ type KeyData = {
 }
 
 type ExportType =
+    | 'types'
     | 'require'
     | 'import'
     | 'default'
     | 'browser_default'
+    | 'browser_types'
     | 'browser_import'
     | 'browser_umd'
 const export_key_lookup: Record<ExportType, KeyData> = {
@@ -60,15 +62,20 @@ const export_key_lookup: Record<ExportType, KeyData> = {
         internal_format: 'es',
         module_format: 'esm',
     },
-    /* types: {
-        extension: 'd.ts',
-        internal_format: "ts",
-        module_format: "typescript"
-    },*/
+    types: {
+        extension: '.d.ts',
+        internal_format: 'es',
+        module_format: 'typescript',
+    },
     browser_default: {
         extension: '-iife.js',
         internal_format: 'iife',
         module_format: 'iife',
+    },
+    browser_types: {
+        extension: '.d.ts',
+        internal_format: 'es',
+        module_format: 'typescript',
     },
     browser_import: {
         extension: '.js',
@@ -94,7 +101,9 @@ const getOutfileName = (
 const getExportKey = (_export_key: string) => {
     return _export_key === '.' || _export_key === '*' || _export_key === 'main'
         ? '.'
-        : _export_key
+        : _export_key.charAt(0) !== '.'
+          ? `./${_export_key}`
+          : _export_key
 }
 const addMinFileExtension = (_value: string, insert_value: string = '.min') => {
     const insert = (
@@ -162,7 +171,7 @@ const getOutputObj = (
         export_type: ExportType
         file: string
         format: InternalModuleFormat
-        export_key?: string
+        export_key: string
     }
     //return minimal objects so we can get an export map later
     const expandedExportTypes: ExpandedExportType[] = entry.export_types.map(
@@ -173,10 +182,12 @@ const getOutputObj = (
             const file = path.resolve(`${output_dir}/${filename}${extension}`)
             const format: InternalModuleFormat =
                 export_key_lookup[export_type].internal_format
+            const export_key = entry.export_key
             return {
                 export_type,
                 file,
                 format,
+                export_key,
             }
         },
     )
@@ -187,7 +198,7 @@ const getOutputObj = (
         >((acc, value: ExpandedExportType) => {
             return {
                 ...acc,
-                [value.export_type]: path.relative('.', value.file), //processTranscriptionSlice(value)
+                [value.export_type]: `./${path.relative('.', value.file)}`, //processTranscriptionSlice(value)
             }
         }, {}),
     }
@@ -218,7 +229,16 @@ const getOutputObj = (
                     : []),
                 mainOutputObject,
             ]
-            return [...acc, ...newOutputArray]
+            console.log(
+                'regexp',
+                new RegExp(/types/, 'g').test(value.export_type),
+            )
+            return [
+                ...acc,
+                ...(new RegExp(/types/, 'g').test(value.export_type) === false
+                    ? newOutputArray
+                    : []),
+            ]
         }, []),
     }
     return { exportObj: unflatten(export_object, { delimiter: '_' }), config }
@@ -229,7 +249,6 @@ const getOutputObj = (
 
 const CDN_PLUGINS_LIST = [
     ts({
-        browserslist: false,
         tsconfig: (resolvedConfig) => ({
             ...resolvedConfig,
             declaration: true,
@@ -247,12 +266,15 @@ const CDN_PLUGINS_LIST = [
 const RESOLVED_PLUGINS_LIST = [
     ts({
         browserslist: false,
-        tsconfig: (resolvedConfig) => ({
-            ...resolvedConfig,
-            declaration: true,
-            allowJs: false,
-            sourceMap: true,
-        }) /* Plugin options */,
+        tsconfig: (resolvedConfig) => {
+            //  console.log("THE RESOLVRDF CONGIG ", resolvedConfig);
+            return {
+                ...resolvedConfig,
+                declaration: true,
+                allowJs: false,
+                sourceMap: true,
+            }
+        } /* Plugin options */,
     }),
     json(),
     nodePolyfills(),
@@ -271,7 +293,7 @@ const outputObjectsArr = [
         ...getOutputObj({
             source_dir: './src/',
             output_dir: './dist/',
-            export_types: ['default', 'import', 'require'],
+            export_types: ['default', 'import', 'require', 'types'],
             export_key: '*',
             library_name: 'gLibrary',
         }),
@@ -280,7 +302,7 @@ const outputObjectsArr = [
     {
         ...getOutputObj({
             ...getDirectoryObj,
-            export_types: ['default', 'import', 'require'],
+            export_types: ['default', 'import', 'require', 'types'],
             export_key: 'node',
             library_name: 'gLibraryNode',
         }),
@@ -292,8 +314,8 @@ const outputObjectsArr = [
             export_types: [
                 'browser_import',
                 'browser_default',
-                'default',
                 'browser_umd',
+                'browser_types',
             ],
             out_file_name_override: 'cdn-index',
             export_key: '*',
@@ -317,5 +339,5 @@ const result = outputObjectsArr.reduce<Record<string, {}>>((acc, value) => {
     const obj = value.exportObj
     return deepmerge(acc, value.exportObj)
 }, {})
-console.log('IMPORTS OBJECT', JSON.stringify(result))
+console.log('IMPORTS OBJECT', JSON.stringify(result, undefined, 4))
 export default CONFIG