+++ packages/g-library/src/string/index.ts
@@ -24,18 +24,15 @@ export {
 
 import * as StringUtils from './_stringUtils.js'
 import * as StringGen from './_string.js'
-import * as Flat from 'flat'
 import * as Markdown from './_markdown.js'
 import * as stringTransform from './../transformString/index.js'
 
 export const stringUtils = {
     ...StringUtils,
     ...StringGen,
-    ...Flat,
     ...Markdown,
 
     ...stringTransform,
-    // ...stringTransform
 }
 
 export type { URL } from './_string.js'